Support selecting schemes and versions
Adding support for selecting different schemes using the option `-s`
when doing benchmarking.

Each scheme is represented as a file with functions that are called to
perform each step in the benchmarking process. Schemes can have
different tables for storing the received data.

It also add options to pick a specific version to use when running the
measurements using the `-v` flag. If no version is picked, the default
version will be used.

Also add support for connecting to different ports.